explicitly define types for {Server,Client}Event
this is forwards-compatible as long as the server is
backwards-compatible; `ClientEvent` should always be accepted on their
side unless there's a breaking API change, and for `ServerEvent`
TypeScript doesn't do any runtime checking, so as long as the values we
expect are still there it'll run great.

curiously, this is slightly inconsistent with the API as defined in the
PDF; item_added.message_id is actually item_added.id in the impl.