chore(agents): explicitly unsupport CommonJS
there's a weird edge-case with `tsx` where if `module` is unset in
package.json but the typescript file still uses module imports, it will
still run, but as CommonJS, without checking for inconsistencies. this
doesn't immediately break agents, instead erroring on `job_main.js`,
without throwing a useful error.

we do not support and have not supported CommonJS, this PR just throws
an error for this specific usecase which has thrown a few folks for a
loop.