fix(elevenlabs): fix errors capturing frames
the error seems to have been with Node.js array buffer handling. copying
the implementation from the newer Cartesia plugin seems to work wonders,
though i'm unsure if that's because it's fixed or because i couldn't
come across the issue; it's a weird one to reproduce.

drive-by removed flake.* and .envrc as i don't use nix anymore